move components Footer, About, Header, Main to folders and change thrm to tsx
