+++ src/App.css
@@ -1,5 +1,3 @@
-@import "./index.css";
-
 :root {
   --dark-green: #166563;
   --medium-green: #93bcbb;
