+++ src/App.js
@@ -1,10 +1,10 @@
-import "./styles/App.css";
+import "./App.css";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Recipes from "./components/Recipes/Recipes.tsx";
-import About from "./components/AboutPage";
-import Contact from "./components/ContactPage";
+import About from "./components/About/About.tsx";
+import Contact from "./components/Contact/Contact.tsx";
 import RecipeDetails from "./components/Recipes/RecipeDetails/RecipeDetails.tsx";
-import MainContent from "./components/MainContent";
+import Main from "./components/Main/Main.tsx";
 import Home from "./components/Home/Home.tsx";
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
     <Router>
       <div className="App">
         <Routes>
-          <Route path="/" element={<MainContent />}>
+          <Route path="/" element={<Main />}>
             <Route path="/" element={<Home />} />
             <Route path="/recipes" element={<Recipes />} />
             <Route path="/about" element={<About />} />
