+++ src/components/About/About.css
@@ -1,5 +1,3 @@
-@import "./App.css";
-
 .about-content {
   margin-top: var(--header-hight);
   padding: 20px;
