--- src/components/AboutPage.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import Header from "./Header";
-import AboutContent from "./AboutContent";
-import Footer from "./Footer";
-
-function About() {
-  return (
-    <div className="about">
-      <Header />
-      <AboutContent />
-      <Footer />
-    </div>
-  );
-}
-
-export default About;
