+++ src/components/Contact/Contact.css
@@ -1,5 +1,3 @@
-@import "./App.css";
-
 :root {
   --modal-width: 500px;
   --modal-height: 380px;
