--- src/components/ContactPage.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-import Header from "./Header";
-import ContactContent from "./ContactContent";
-import Footer from "./Footer";
-
-function Contact() {
-  return (
-    <div className="contact">
-      <Header />
-      <ContactContent />
-      <Footer />
-    </div>
-  );
-}
-
-export default Contact;
