+++ src/components/Footer/Footer.css
@@ -1,5 +1,3 @@
-@import "./App.css";
-
 :root {
   --footer-hight: 180px;
   --footer-width: 100%;
