+++ src/components/Header/Header.css
@@ -1,5 +1,3 @@
-@import "./App.css";
-
 :root {
   --header-hight: 120px;
   --header-width: 100%;
