+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import "../styles/Header.css";
+import "./Header.css";
 
 function Header() {
   return (
