+++ src/components/Main/Main.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import Header from "../Header/Header";
+import Footer from "../Footer/Footer";
+import { Outlet } from "react-router-dom";
+
+function Main() {
+  return (
+    <>
+      <Header />
+      <Outlet />
+      <Footer />
+    </>
+  );
+}
+
+export default Main;
