+++ src/index.css
@@ -1,0 +1,6 @@
+body {
+  font-family: Nunito, sans-serif;
+  margin: 0;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
